Set up `modules` directory and move functions around
