+++ functions/tsconfig.json
@@ -9,7 +9,8 @@
     "strict": true,
     "target": "es2021",
     "resolveJsonModule": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "esModuleInterop": true
   },
   "compileOnSave": true,
   "include": ["src/**/*.ts", "../src/models/Event.ts"],
